fix) 오답노트 페이지 버튼 구조 변경
