+++ src/pages/NotePage/ConceptExplanation.jsx
@@ -1,17 +1,9 @@
 import React from "react";
 import styled from "styled-components";
-import Button from "../../components/SolveButton";
 import DownloadButton from "../../components/DownloadButton";
 import { conceptData } from "./data/conceptData";
-import { useNavigate } from "react-router-dom";
 
 const ConceptExplanation = () => {
-  const navigate = useNavigate();
-
-  const handleButtonClick = () => {
-    navigate("/Addcomplete");
-  };
-
   return (
     <Container>
       <RightSection>
@@ -45,12 +37,6 @@ const ConceptExplanation = () => {
           ))}
         </ConceptsWrapper>
       </RightSection>
-      <SolveButtonWrapper>
-        <Button variant="filled" onClick={handleButtonClick}>
-          한숨 쉴 시간에 한 문제라도 더 풀자 <br />
-          추가 연습 문제 생성하기
-        </Button>
-      </SolveButtonWrapper>
     </Container>
   );
 };
@@ -58,12 +44,6 @@ const ConceptExplanation = () => {
 const Container = styled.div`
   display: flex;
   flex-direction: column;
-`;
-
-const SolveButtonWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  margin: 50px 0;
 `;
 
 const RightSection = styled.div`
