+++ src/pages/NotePage/WrongProblemList.jsx
@@ -1,15 +1,10 @@
 import React from "react";
 import styled from "styled-components";
-import { useNavigate } from "react-router-dom";
 import MultipleChoice from "../PracticePage/MultipleChoice";
-import Button from "../../components/SolveButton";
 
 const WrongProblemList = ({ problems }) => {
-  const navigate = useNavigate();
 
-  const handleButtonClick = () => {
-    navigate("/mypage/summary");
-  };
+
 
   const renderProblem = (problem) => {
     if (problem.type === "multiple_choice") {
@@ -40,13 +35,6 @@ const WrongProblemList = ({ problems }) => {
   return (
     <Container>
       <ProblemListContainer>{problems.map(renderProblem)}</ProblemListContainer>
-      <ButtonWrapper>
-        <Button onClick={handleButtonClick}>
-          많이 틀렸어도 기죽지 말자! 앞으로도 화이팅!
-          <br />
-          마이페이지로 이동하기
-        </Button>
-      </ButtonWrapper>
     </Container>
   );
 };
@@ -136,10 +124,5 @@ const Container = styled.div`
   margin-top: 100px;
 `;
 
-const ButtonWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  margin: 50px 0;
-`;
 
 export default WrongProblemList;
